% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_mbs.R
\name{plotly_mbs}
\alias{plotly_mbs}
\title{MBS Energy Plotly Theme}
\usage{
plotly_mbs(
  plot,
  ...,
  plot_title = "Insert Plot's Title",
  plot_subtitle = "",
  plot_caption = "",
  plot_xaxis = "",
  plot_yaxis = ""
)
}
\arguments{
\item{plot}{plotly plot object}

\item{...}{inherit arguments coming from the plotly's layout function}

\item{plot_title}{Main title}

\item{plot_subtitle}{Subtitle in default position}

\item{plot_caption}{Caption in default position}

\item{plot_xaxis}{x-axis text}

\item{plot_yaxis}{y-axis text}
}
\description{
MBS Energy Plotly Theme
}
\examples{
library(plotly)

df = data.table(State = state.name,
                LifeExp = state.x77[,4],
                Population = state.x77[,1])

df |>
 plot_ly(x = ~LifeExp, y = ~Population, type = "scatter", width = 900, height = 500) |>
   plotly_mbs(plot_title = 'This is my title',
                  plot_subtitle = 'Here we will see a subtitle',
                  plot_caption = 'Source: innteamUtils package',
                  plot_xaxis = 'LifeExp',
                  plot_yaxis = 'Population')

}
