% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mbslight.R
\name{theme_mbslight}
\alias{theme_mbslight}
\title{MBS ggpolt2 Theme}
\usage{
theme_mbslight(base_size = 11)
}
\arguments{
\item{base_size}{base font size, given in pts.}
}
\description{
Contains the parameters and colour for a standarized plot theme.
}
\author{
Alejandro Abraham \href{mailto:a.abraham@mbsconsulting.it}{a.abraham@mbsconsulting.it}
}
